Unify TerrainCfg into single path and remove duplicate entities config
